refactor finfet tests;minor fixes
