+++ tests/pdk/finfet_pdk/test_generator.py
@@ -1,8 +1,5 @@
 from align.primitive import main
-try:
-    from .utils import pdk_dir, export_to_viewer
-except ImportError:
-    from utils import pdk_dir, export_to_viewer
+from .utils import pdk_dir, export_to_viewer
 
 
 def test_example():
